# Optimized Actions API