Optimized the Actions API by:

1. Removing the encode/decode process when storing/getting actions from the container. Closes #112.
     - Before change: 
          - `_beans_set_action` was 0.012817ms slower than Beans 1.4.0
          - `_beans_get_action` was 0.00499ms faster than Beans 1.4.0
          - The set function impacted multiple functions.
     - After change: 
          - `_beans_set_action` is 0.001123ms _faster_ than Beans 1.4.0
          - All functions that use set or get are now all faster than Beans 1.4.0
2. Shifting the emphasis to the developer to ensure s/he knows how to add an action. 
     - This eliminated checking that each parameter is set and not `null`. 
     - Removed `_beans_is_action_valid()` function.
3. Rethought `_beans_get_current_action()` - it seems highly unlikely that a developer would need to modify all 4 action parameters (i.e. hook, callback, priority, and args).  Therefore, we can bail out if there is no "added" action registered.  
    - Before change: 0.000619ms slower than Beans 1.4.0. 
    - Now 0.003675ms _faster_ than Beans 1.4.0.
4. Standardized the action validity check to use `empty()`.

## Micro Profiler Report BEFORE the above optimizations:

```
-------------------------
MICRO PROFILER REPORT:
-------------------------

PHP version:        5.6.20
Sample Size:        10,000 (per function)
Time Increments:    milliseconds, where 1 ms equals 0.001 second.

  --------------------------------   -------------   -------------   -------------
| Name                             | Result        |  Avg Time     | v1.4.0
|                                  | (ms)          |  (ms)         | (ms)
| -------------------------------- | ------------- | ------------- | -------------
| _beans_get_action                |    -0.000499  |      0.009867 |      0.010366
| _beans_set_action                |     0.012817  |      0.017634 |      0.004817
| _beans_unset_action              |    -0.000452  |      0.009037 |      0.009490
| _beans_merge_action              |     0.010362  |      0.024231 |      0.013869
| _beans_get_current_action        |     0.000619  |      0.016486 |      0.015867
| beans_add_smart_action           |     0.011813  |      0.033574 |      0.021761
| beans_modify_action_priority     |     0.013136  |      0.054333 |      0.041198
| beans_replace_action_callback    |     0.035778  |      0.097265 |      0.061487
| beans_remove_action              |     0.001023  |      0.016903 |      0.015880
| beans_reset_action               |     0.003364  |      0.044388 |      0.041024
 --------------------------------   -------------   -------------   -------------

NOTES:

1. The functions were invoked/exercised 10,000 times during this micro profiler process.
2. The 'diff' column = 'time' - 'v1.4.0'.
3. The 'time' column shows the function's average execution time.
4. 'v1.4.0' shows the same micro-profiler run on Beans v1.4.0.
```

## Report AFTER the Optimizations

```
-------------------------
MICRO PROFILER REPORT:
-------------------------

PHP version:        5.6.20
Sample Size:        10,000 (per function)
Time Increments:    milliseconds, where 1 ms equals 0.001 second.

  --------------------------------   -------------   -------------   -------------
| Name                             | Result        |  Avg Time     | v1.4.0
|                                  | (ms)          |  (ms)         | (ms)
| -------------------------------- | ------------- | ------------- | -------------
| _beans_get_action                |    -0.003312  |      0.007054 |      0.010366
| _beans_set_action                |    -0.000801  |      0.004016 |      0.004817
| _beans_unset_action              |    -0.002413  |      0.007077 |      0.009490
| _beans_merge_action              |    -0.003856  |      0.010013 |      0.013869
| _beans_get_current_action        |    -0.003298  |      0.012569 |      0.015867
| beans_add_smart_action           |    -0.001909  |      0.019853 |      0.021761
| beans_modify_action_priority     |    -0.005301  |      0.035896 |      0.041198
| beans_replace_action_callback    |    -0.008353  |      0.053134 |      0.061487
| beans_remove_action              |    -0.004880  |      0.011001 |      0.015880
| beans_reset_action               |    -0.006546  |      0.034478 |      0.041024
 --------------------------------   -------------   -------------   -------------

NOTES:

1. The functions were invoked/exercised 10,000 times during this micro profiler process.
2. The 'diff' column = 'time' - 'v1.4.0'.
3. The 'time' column shows the function's average execution time.
4. 'v1.4.0' shows the same micro-profiler run on Beans v1.4.0.
```