@hellofromtonya 2018-02-12
To measure the execution times, I used a [Micro Profiler plugin](https://github.com/hellofromtonya/beans-profiler) that I'm developing for us.  It does the following:

1. Cycles each configured function (profile) 10,000 times.
2. Averages all of the executive times and reports it in milliseconds.
3. It then compares the average to the baseline average (which right now it's comparing to Beans 1.4.0).

This tool is a bit crude at the moment, but does the job we need right now for measuring actions and filters.